requirements.txt
flask
face_recognition
pillow
